Fix test breaks on KitKat or prior devices
This CL splits MozcService into MozcService and MozcServiceImpl for
testing.  MozcService converts CursorAnchorInfo to home-made
CursorAnchorInfoWrapper, and pass it to MozcServiceImpl. It allows
EasyMock to mocking MozcServiceImpl.  Otherwise, EasyMock raises
NoClassDefFoundError for CursorAnchorInfo on API 20.

BUG=
TEST=compile
REF_BUG=18779687
REF_CL=83582400
